throw a better exception if requested CF not found
If someone tries to import data to a cassandra column family that
doesn't actually exist (maybe they forgot to create the table), the
Crunch job fails with a NullPointerException at the
`tableMetadata.getColumns()` line since there is no metadata for the
table. Throw a more descriptive error message in this case.