CqlRecord creation with columns keyed by their names
It was reported to be difficult to check the ordering of columns
in table's CQL schema, and then compose a list of values passed to
CqlRecord.create() in a wat it matches the CQL schema. Moreover,
this sometimes failed to correspond to the order expected by the
prepared insert statement used by the output writer.

This change allows creating CqlRecords by passing maps of columns keyed by their names.
To allow this, hdfs2cass must internally store the ordering of columns in the insert
statement. It's also needed to place this information in job's config so that it gets
propagated to workers' processes.