Infer partitioning keys from CQL records
When constructing a CQLRecord from a binary key and column values, it
is easy to get the partitioning key wrong (for example by forgetting
to include some components, or getting the encoding wrong). This
throws the partitioner off, and causes records to be sent to
essentially random reducers. We end up with all reducers streaming to
all Cassandra nodes, which works but is very undesirable (creating way
too many sstables on each node).

Fortunately, the CQL record together with the schema contain all
information needed to build the partitioning key. This patch lets the
framework calculate the partitioning key, simplifying user code.